requirements.txt
Django==1.10.4
django-braces==1.10.0
django-cors-headers==1.3.1
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
oauthlib==1.0.3
psycopg2==2.6.2
pymongo==3.4.0
pytz==2016.10
requests==2.12.4
pbr==1.10.0
python-Levenshtein==0.12.0
six==1.10.0
stevedore==1.19.1
pyjwkest==1.3.2
